Mudado para Long o tipo do parâmetro de deletaLocal()

+++ app/src/main/java/com/infact/nightour/controller/LocaisController.java
@@ -81,7 +81,7 @@ public class LocaisController {
         db.close();
     }
 
-    public void deletaLocal(int id) {
+    public void deletaLocal(Long id) {
         String where = Local.BD_ID + " = " + id;
 
         db = banco.getWritableDatabase();
